+++ lib/src/wire/builders/widget_builder.dart
@@ -141,3 +141,4 @@ part 'widget_builders/ping_error_builder.dart';
 part 'widget_builders/center_builder.dart';
 
 part 'widget_builders/choice_chip_builder.dart';
+part 'widget_builders/flexible_builder.dart';